test(): move replay sig fn to aztec.js
