+++ packages/aztec.js/src/signer/index.js
@@ -67,6 +67,24 @@ signer.generateDAIDomainParams = (chainId, verifyingContract) => {
     };
 };
 
+signer.makeReplaySignature = (signatureToReplay) => {
+    const [r, s, v] = signatureToReplay
+        .slice(2)
+        .match(/.{1,64}/g);
+    const secp256k1n = new BN("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEBAAEDCE6AF48A03BBFD25E8CD0364141", 16);
+    const hex28 = (new BN(28)).toString(16);
+    const hex27 = (new BN(27)).toString(16);
+
+    const flippedS = secp256k1n
+        .sub(new BN(s, 16))
+        .toString(16);
+    const flippedV = (v === '1b') ? padRight(hex28.slice(-2), 64) : padRight(hex27.slice(-2), 64);
+
+    const reconstructedSig = r + flippedS + flippedV;
+    return `0x${reconstructedSig.slice(0, 130)}`
+}
+
+
 /**
  * Create an EIP712 ECDSA signature over an AZTEC note, suited for the confidentialApprove() method of a
  * ZkAsset. The ZkAsset.confidentialApprove() method must be called when granting note spending permission
@@ -86,7 +104,7 @@ signer.generateDAIDomainParams = (chainId, verifyingContract) => {
  * @param {string} privateKey the private key of message signer
  * @returns {string} ECDSA signature parameters [r, s, v], formatted as 32-byte wide hex-strings
  */
-signer.signNoteForConfidentialApprove = (verifyingContract, noteHash, spender, spenderApproval, privateKey) => {
+signer.signNoteForConfidentialApprove = (verifyingContract, noteHash, spender, spenderApproval, privateKey, flip = false) => {
     if (verifyingContract === '0x7dd4e19395c47753370a7e20b3788546958b2ea6') {
         console.warn('The signature you are generating can be replayed once on this zkAsset');
         console.warn('To avoid unexpected behaviour, submit the signature returned by this function and');
@@ -103,6 +121,9 @@ signer.signNoteForConfidentialApprove = (verifyingContract, noteHash, spender, s
 
     const { unformattedSignature } = signer.signTypedData(domain, schema, message, privateKey);
     const signature = `0x${unformattedSignature.slice(0, 130)}`; // extract r, s, v (v is just 1 byte, 2 characters)
+    if (flip) {
+        return signer.makeReplaySignature(signature);
+    }
     return signature;
 };
 
@@ -122,7 +143,7 @@ signer.signNoteForConfidentialApprove = (verifyingContract, noteHash, spender, s
  * @param {string} privateKey the private key of message signer
  * @returns {string} ECDSA signature parameters [r, s, v], formatted as 32-byte wide hex-strings
  */
-signer.signApprovalForProof = (verifyingContract, proofOutputs, spender, approval, privateKey) => {
+signer.signApprovalForProof = (verifyingContract, proofOutputs, spender, approval, privateKey, flip = false) => {
     if (verifyingContract === '0x7dd4e19395c47753370a7e20b3788546958b2ea6') {
         console.warn('The signature you are generating can be replayed once on this zkAsset');
         console.warn('To avoid unexpected behaviour, submit the signature returned by this function and');
@@ -140,6 +161,9 @@ signer.signApprovalForProof = (verifyingContract, proofOutputs, spender, approva
     };
     const { unformattedSignature } = signer.signTypedData(domain, schema, message, privateKey);
     const signature = `0x${unformattedSignature.slice(0, 130)}`; // extract r, s, v (v is just 1 byte, 2 characters)
+    if (flip) {
+        return signer.makeReplaySignature(signature);
+    }
     return signature;
 };
